Update bitflags dependency version from 0.5 to 0.7
This new version requires a Rust compiler version of at least 1.8 (from
April 2016). Previously we've only required 1.4 or better. The bitflags
dependency is optional, and so this should only affect users of the
'xmp-packet-access' feature, but for simplicity in documentation we'll
just say that that's the earliest version we support in general.

Fixes #14.